improved logic to find the best route
